move zones resources to their own file
